move test case subscription to editable instructions
