+++ src/main/webapp/app/entities/programming-exercise/instructions/programming-exercise-instruction.component.ts
@@ -17,8 +17,7 @@ import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { HttpResponse } from '@angular/common/http';
 import { TranslateService } from '@ngx-translate/core';
 import * as Remarkable from 'remarkable';
-import { intersection as _intersection } from 'lodash';
-import { faCheckCircle, faTimesCircle, faQuestionCircle } from '@fortawesome/free-regular-svg-icons';
+import { faCheckCircle, faQuestionCircle, faTimesCircle } from '@fortawesome/free-regular-svg-icons';
 import { catchError, filter, flatMap, map, switchMap, tap } from 'rxjs/operators';
 import { CodeEditorService } from 'app/code-editor/service/code-editor.service';
 import { EditorInstructionsResultDetailComponent } from 'app/code-editor/instructions/code-editor-instructions-result-detail';
@@ -26,12 +25,11 @@ import { Feedback } from 'app/entities/feedback';
 import { Result, ResultService } from 'app/entities/result';
 import { ProgrammingExercise } from '../programming-exercise.model';
 import { RepositoryFileService } from 'app/entities/repository';
-import { Participation, hasParticipationChanged, ParticipationWebsocketService } from 'app/entities/participation';
+import { hasParticipationChanged, Participation, ParticipationWebsocketService } from 'app/entities/participation';
 import { FaIconComponent } from '@fortawesome/angular-fontawesome';
 import { Observable, Subscription } from 'rxjs';
 import { hasExerciseChanged, problemStatementHasChanged } from 'app/entities/exercise';
 import { ProgrammingExerciseTestCase } from 'app/entities/programming-exercise/programming-exercise-test-case.model';
-import { ProgrammingExerciseTestCaseService } from 'app/entities/programming-exercise/services';
 import { isLegacyResult } from 'app/entities/programming-exercise/utils/programming-exercise.utils';
 
 export enum TestCaseState {
@@ -67,8 +65,6 @@ export class ProgrammingExerciseInstructionComponent implements OnChanges, OnDes
     public resultChange = new EventEmitter<Result>();
     @Output() public exerciseTestCasesChange = new EventEmitter<ProgrammingExerciseTestCase[] | null>();
 
-    exerciseTestCases: string[] | null = null;
-
     public problemStatement: string;
     public participationSubscription: Subscription;
 
@@ -81,15 +77,12 @@ export class ProgrammingExerciseInstructionComponent implements OnChanges, OnDes
     // Can be used to remove the click listeners for result details
     private listenerRemoveFunctions: Function[] = [];
 
-    testCaseSubscription: Subscription;
-
     constructor(
         private editorService: CodeEditorService,
         private translateService: TranslateService,
         private resultService: ResultService,
         private repositoryFileService: RepositoryFileService,
         private participationWebsocketService: ParticipationWebsocketService,
-        private testCaseService: ProgrammingExerciseTestCaseService,
         private renderer: Renderer2,
         private elementRef: ElementRef,
         private modalService: NgbModal,
@@ -114,9 +107,6 @@ export class ProgrammingExerciseInstructionComponent implements OnChanges, OnDes
         const participationHasChanged = hasParticipationChanged(changes);
         const exerciseHasChanged = hasExerciseChanged(changes);
         // It is possible that the exercise does not have an id in case it is being created now.
-        if (exerciseHasChanged && this.exercise.id) {
-            this.setupTestCaseSubscription();
-        }
         if (participationHasChanged) {
             this.isInitial = true;
             this.setupResultWebsocket();
@@ -152,25 +142,6 @@ export class ProgrammingExerciseInstructionComponent implements OnChanges, OnDes
             this.problemStatement = this.exercise.problemStatement!;
             this.updateMarkdown();
         }
-    }
-
-    private setupTestCaseSubscription() {
-        if (this.testCaseSubscription) {
-            this.testCaseSubscription.unsubscribe();
-        }
-
-        this.testCaseSubscription = this.testCaseService
-            .subscribeForTestCases(this.exercise.id)
-            .pipe(
-                tap(testCases => this.exerciseTestCasesChange.emit(testCases)),
-                filter(testCases => !!testCases),
-                tap(testCases => {
-                    this.exerciseTestCases = testCases && testCases.filter(({ active }) => active).map(({ testName }) => testName);
-                }),
-                // The test cases validate the task specific tests, so we need to update the markdown here.
-                tap(() => this.updateMarkdown()),
-            )
-            .subscribe();
     }
 
     /**
@@ -403,12 +374,11 @@ export class ProgrammingExerciseInstructionComponent implements OnChanges, OnDes
             // In silent mode it shouldn't output any tokens or modify pending
             if (!silent) {
                 const tests = match[2].split(',');
-                const validTests = this.exerciseTestCases && this.latestResult && !isLegacyResult(this.latestResult) ? _intersection(tests, this.exerciseTestCases) : tests;
                 // Insert the testsStatus token to our rendered tokens
                 state.push({
                     type: 'testsStatus',
                     title: match[1],
-                    tests: validTests,
+                    tests,
                     level: state.level,
                 });
             }
@@ -506,17 +476,16 @@ export class ProgrammingExerciseInstructionComponent implements OnChanges, OnDes
         const tests = tokens[0].tests || [];
         const [done, label] = this.statusForTests(tests);
         const textColor = done === TestCaseState.SUCCESS ? 'text-success' : done === TestCaseState.FAIL ? 'text-danger' : 'text-secondary';
-        const validTestCases = this.exerciseTestCases && this.latestResult && !isLegacyResult(this.latestResult) ? _intersection(tests, this.exerciseTestCases).toString() : tests;
 
         let text = `<span class="bold"><span id=step-icon-${this.steps.length}></span>`;
 
         text += ' ' + tokens[0].title;
         text += '</span>: ';
         // If the test is not done, we set the 'data-tests' attribute to the a-element, which we later use for the details dialog
-        if (done === TestCaseState.SUCCESS || done === TestCaseState.NO_RESULT || !validTestCases.length) {
+        if (done === TestCaseState.SUCCESS || done === TestCaseState.NO_RESULT || !tests.length) {
             text += `<span class="${textColor} bold">` + label + '</span>';
         } else if (done === TestCaseState.FAIL || done === TestCaseState.NOT_EXECUTED) {
-            text += '<a data-tests="' + validTestCases + `" class="test-status"><span class="${textColor} result">` + label + '</span></a>';
+            text += '<a data-tests="' + tests + `" class="test-status"><span class="${textColor} result">` + label + '</span></a>';
         }
         text += '<br>';
 
@@ -609,9 +578,6 @@ export class ProgrammingExerciseInstructionComponent implements OnChanges, OnDes
         this.steps = [];
         if (this.participationSubscription) {
             this.participationSubscription.unsubscribe();
-        }
-        if (this.testCaseSubscription) {
-            this.testCaseSubscription.unsubscribe();
         }
     }
 }
